chore: DIAG-3, LEX-5, SRCMGR-2 lexer/diagnostics consistency sweep
Three small findings in the same subsystem, none of which changes behavior.

DIAG-3 -- DiagnosticEngine::report checks the MAX_ERRORS cap before examining
the level, so once 20 errors are reached warnings are dropped too, not just
errors. Confirmed deliberate and documented as such: past the cap the
compilation is already abandoned and later stages are reasoning about a broken
program, so their warnings are as likely to be noise as signal -- and emitting
anything after "too many errors emitted, stopping." reads as if it had not
stopped.

LEX-5 -- line_/col_ (and the token-start snapshot added by LEX-2) were
uint32_t, while the SourceLocation fields they feed are size_t, and
asm_lexer.cpp tracks the same quantity as size_t. Widened to size_t so one
concept has one width across both lexers. No truncation was possible in
practice, but the mismatch invited exactly the kind of narrowing arithmetic
that produced LEX-2.

SRCMGR-2 -- get_source_line built a temporary std::string from its
string_view argument on every call, purely to probe
unordered_map<std::string, std::string>. Since it is called once per
diagnostic printed, that is an allocation per diagnostic. Added a transparent
hasher/comparator so the string_view is looked up directly.

+++ src/compiler/lexer.cpp
@@ -121,8 +121,8 @@ namespace lexer {
             std::string_view filename_;
             DiagnosticEngine &diagnostics_;
             size_t pos_ = 0;
-            uint32_t line_ = 1;
-            uint32_t col_ = 1;
+            size_t line_ = 1;
+            size_t col_ = 1;
             std::optional<TokenKind> last_real_kind_ = std::nullopt;
             // Set right after a 'KwAsm' token whose immediate lex_asm_block() attempt failed
             // (i.e. the next real char is '-' '>' — the expression form's header, not the
@@ -145,8 +145,8 @@ namespace lexer {
             // back-computing the start from the end position, which is only correct while a
             // token stays on one line. lex_asm_block() has always snapshotted its own start
             // for exactly this reason; these two members generalize that to every token.
-            uint32_t token_start_line_ = 1;
-            uint32_t token_start_col_ = 1;
+            size_t token_start_line_ = 1;
+            size_t token_start_col_ = 1;
 
             LexerImpl(const std::string_view source, const std::string_view filename, DiagnosticEngine &diagnostics) : source_(source), filename_(filename), diagnostics_(diagnostics) {}
